Retry transactions that fail due to serializable isolation level (#736)
* split up database/mod.rs

* retry

* retry macro rules

* added warning for failed tx commits
