+++ Cargo.toml
@@ -93,7 +93,7 @@ semaphore = { git = "https://github.com/worldcoin/semaphore-rs", rev = "5170c422
 similar-asserts = "1.5.0"
 test-case = "3.0"
 testcontainers = "0.15.0"
-testcontainers-modules = {  version = "0.3.7", features = ["postgres"] }
+testcontainers-modules = { version = "0.3.7", features = ["postgres"] }
 tracing-subscriber = "0.3.11"
 tracing-test = "0.2"
