+++ src/database/mod.rs
@@ -5,31 +5,25 @@
 )]
 
 use std::cmp::Ordering;
-use std::collections::HashSet;
 use std::ops::Deref;
 
 use anyhow::{anyhow, Context, Error as ErrReport};
-use chrono::{DateTime, Utc};
-use ruint::aliases::U256;
 use sqlx::migrate::{Migrate, MigrateDatabase, Migrator};
 use sqlx::pool::PoolOptions;
 use sqlx::{Executor, Pool, Postgres, Row};
 use thiserror::Error;
 use tracing::{error, info, instrument, warn};
 
-use self::types::{CommitmentHistoryEntry, DeletionEntry, LatestDeletionEntry, RecoveryEntry};
 use crate::config::DatabaseConfig;
-use crate::identity_tree::{
-    Hash, ProcessedStatus, RootItem, TreeItem, TreeUpdate, UnprocessedStatus,
-};
-use crate::prover::{ProverConfig, ProverType};
+use crate::database::query::DatabaseQuery;
+use crate::identity_tree::Hash;
 
+pub mod query;
+pub mod transaction;
 pub mod types;
 
 // Statically link in migration files
 static MIGRATOR: Migrator = sqlx::migrate!("schemas/database");
-
-const MAX_UNPROCESSED_FETCH_COUNT: i64 = 10_000;
 
 pub struct Database {
     pub pool: Pool<Postgres>,
@@ -43,7 +37,7 @@ impl Deref for Database {
     }
 }
 
-impl<'a, T> DatabaseExt<'a> for T where T: Executor<'a, Database = Postgres> {}
+impl<'a, T> DatabaseQuery<'a> for T where T: Executor<'a, Database = Postgres> {}
 
 impl Database {
     #[instrument(skip_all)]
@@ -142,704 +136,6 @@ impl Database {
 
         Ok(Self { pool })
     }
-
-    /// Marks the identities and roots from before a given root hash as mined
-    /// Also marks following roots as pending
-    #[instrument(skip(self), level = "debug")]
-    pub async fn mark_root_as_processed(&self, root: &Hash) -> Result<(), Error> {
-        let mut tx = self.pool.begin().await?;
-
-        let root_id = tx.get_id_by_root(root).await?;
-
-        let Some(root_id) = root_id else {
-            return Err(Error::MissingRoot { root: *root });
-        };
-
-        let root_id = root_id as i64;
-        // TODO: Can I get rid of line `AND    status <> $2
-        let update_previous_roots = sqlx::query(
-            r#"
-            UPDATE identities
-            SET    status = $2, mined_at = CURRENT_TIMESTAMP
-            WHERE  id <= $1
-            AND    status <> $2
-            AND    status <> $3;
-            "#,
-        )
-        .bind(root_id)
-        .bind(<&str>::from(ProcessedStatus::Processed))
-        .bind(<&str>::from(ProcessedStatus::Mined));
-
-        let update_next_roots = sqlx::query(
-            r#"
-            UPDATE identities
-            SET    status = $2, mined_at = NULL
-            WHERE  id > $1
-            "#,
-        )
-        .bind(root_id)
-        .bind(<&str>::from(ProcessedStatus::Pending));
-
-        tx.execute(update_previous_roots).await?;
-        tx.execute(update_next_roots).await?;
-
-        tx.commit().await?;
-
-        Ok(())
-    }
-
-    /// Marks the identities and roots from before a given root hash as
-    /// finalized
-    #[instrument(skip(self), level = "debug")]
-    pub async fn mark_root_as_mined(&self, root: &Hash) -> Result<(), Error> {
-        let mined_status = ProcessedStatus::Mined;
-
-        let mut tx = self.pool.begin().await?;
-        tx.execute("SET TRANSACTION ISOLATION LEVEL SERIALIZABLE;")
-            .await?;
-
-        let root_id = tx.get_id_by_root(root).await?;
-
-        let Some(root_id) = root_id else {
-            return Err(Error::MissingRoot { root: *root });
-        };
-
-        let root_id = root_id as i64;
-
-        let update_previous_roots = sqlx::query(
-            r#"
-            UPDATE identities
-            SET    status = $2
-            WHERE  id <= $1
-            AND    status <> $2
-            "#,
-        )
-        .bind(root_id)
-        .bind(<&str>::from(mined_status));
-
-        tx.execute(update_previous_roots).await?;
-
-        tx.commit().await?;
-
-        Ok(())
-    }
-
-    pub async fn get_identity_history_entries(
-        &self,
-        commitment: &Hash,
-    ) -> Result<Vec<CommitmentHistoryEntry>, Error> {
-        let unprocessed = sqlx::query(
-            r#"
-            SELECT commitment, status, eligibility
-            FROM unprocessed_identities
-            WHERE commitment = $1
-        "#,
-        )
-        .bind(commitment);
-
-        let rows = self.pool.fetch_all(unprocessed).await?;
-        let unprocessed_updates = rows
-            .into_iter()
-            .map(|row| {
-                let eligibility_timestamp: DateTime<Utc> = row.get(2);
-                let held_back = Utc::now() < eligibility_timestamp;
-
-                CommitmentHistoryEntry {
-                    leaf_index: None,
-                    commitment: row.get::<Hash, _>(0),
-                    held_back,
-                    status: row
-                        .get::<&str, _>(1)
-                        .parse()
-                        .expect("Failed to parse unprocessed status"),
-                }
-            })
-            .collect::<Vec<CommitmentHistoryEntry>>();
-
-        let leaf_index = self.get_identity_leaf_index(commitment).await?;
-        let Some(leaf_index) = leaf_index else {
-            return Ok(unprocessed_updates);
-        };
-
-        let identity_deletions = sqlx::query(
-            r#"
-            SELECT commitment
-            FROM deletions
-            WHERE leaf_index = $1
-            "#,
-        )
-        .bind(leaf_index.leaf_index as i64);
-
-        let rows = self.pool.fetch_all(identity_deletions).await?;
-        let deletions = rows
-            .into_iter()
-            .map(|_row| CommitmentHistoryEntry {
-                leaf_index: Some(leaf_index.leaf_index),
-                commitment: Hash::ZERO,
-                held_back:  false,
-                status:     UnprocessedStatus::New.into(),
-            })
-            .collect::<Vec<CommitmentHistoryEntry>>();
-
-        let processed_updates = sqlx::query(
-            r#"
-            SELECT commitment, status
-            FROM identities
-            WHERE leaf_index = $1
-            ORDER BY id ASC
-            "#,
-        )
-        .bind(leaf_index.leaf_index as i64);
-
-        let rows = self.pool.fetch_all(processed_updates).await?;
-        let processed_updates: Vec<CommitmentHistoryEntry> = rows
-            .into_iter()
-            .map(|row| CommitmentHistoryEntry {
-                leaf_index: Some(leaf_index.leaf_index),
-                commitment: row.get::<Hash, _>(0),
-                held_back:  false,
-                status:     row
-                    .get::<&str, _>(1)
-                    .parse()
-                    .expect("Status is unreadable, database is corrupt"),
-            })
-            .collect();
-
-        Ok([processed_updates, unprocessed_updates, deletions]
-            .concat()
-            .into_iter()
-            .collect())
-    }
-}
-
-/// This trait provides the individual and composable queries to the database.
-/// Each method is a single atomic query, and can be composed withing a
-/// transaction.
-pub trait DatabaseExt<'a>: Executor<'a, Database = Postgres> {
-    async fn insert_pending_identity(
-        self,
-        leaf_index: usize,
-        identity: &Hash,
-        root: &Hash,
-    ) -> Result<(), Error> {
-        let insert_pending_identity_query = sqlx::query(
-            r#"
-            INSERT INTO identities (leaf_index, commitment, root, status, pending_as_of)
-            VALUES ($1, $2, $3, $4, CURRENT_TIMESTAMP)
-            "#,
-        )
-        .bind(leaf_index as i64)
-        .bind(identity)
-        .bind(root)
-        .bind(<&str>::from(ProcessedStatus::Pending));
-
-        self.execute(insert_pending_identity_query).await?;
-
-        Ok(())
-    }
-
-    async fn get_id_by_root(self, root: &Hash) -> Result<Option<usize>, Error> {
-        let root_index_query = sqlx::query(
-            r#"
-            SELECT id
-            FROM identities
-            WHERE root = $1
-            ORDER BY id ASC
-            LIMIT 1
-            "#,
-        )
-        .bind(root);
-
-        let row = self.fetch_optional(root_index_query).await?;
-
-        let Some(row) = row else { return Ok(None) };
-        let root_id = row.get::<i64, _>(0);
-
-        Ok(Some(root_id as usize))
-    }
-
-    /// Marks all the identities in the db as
-    #[instrument(skip(self), level = "debug")]
-    async fn mark_all_as_pending(self) -> Result<(), Error> {
-        let pending_status = ProcessedStatus::Pending;
-
-        let update_all_identities = sqlx::query(
-            r#"
-                UPDATE identities
-                SET    status = $1, mined_at = NULL
-                WHERE  status <> $1
-                "#,
-        )
-        .bind(<&str>::from(pending_status));
-
-        self.execute(update_all_identities).await?;
-
-        Ok(())
-    }
-
-    async fn get_next_leaf_index(self) -> Result<usize, Error> {
-        let query = sqlx::query(
-            r#"
-            SELECT leaf_index FROM identities
-            ORDER BY leaf_index DESC
-            LIMIT 1
-            "#,
-        );
-
-        let row = self.fetch_optional(query).await?;
-
-        let Some(row) = row else { return Ok(0) };
-        let leaf_index = row.get::<i64, _>(0);
-
-        Ok((leaf_index + 1) as usize)
-    }
-
-    async fn get_identity_leaf_index(self, identity: &Hash) -> Result<Option<TreeItem>, Error> {
-        let query = sqlx::query(
-            r#"
-            SELECT leaf_index, status
-            FROM identities
-            WHERE commitment = $1
-            ORDER BY id DESC
-            LIMIT 1;
-            "#,
-        )
-        .bind(identity);
-
-        let Some(row) = self.fetch_optional(query).await? else {
-            return Ok(None);
-        };
-
-        let leaf_index = row.get::<i64, _>(0) as usize;
-
-        let status = row
-            .get::<&str, _>(1)
-            .parse()
-            .expect("Status is unreadable, database is corrupt");
-
-        Ok(Some(TreeItem { status, leaf_index }))
-    }
-
-    async fn get_commitments_by_status(
-        self,
-        status: ProcessedStatus,
-    ) -> Result<Vec<TreeUpdate>, Error> {
-        Ok(sqlx::query_as::<_, TreeUpdate>(
-            r#"
-            SELECT leaf_index, commitment as element
-            FROM identities
-            WHERE status = $1
-            ORDER BY id ASC;
-            "#,
-        )
-        .bind(<&str>::from(status))
-        .fetch_all(self)
-        .await?)
-    }
-
-    async fn get_latest_root_by_status(
-        self,
-        status: ProcessedStatus,
-    ) -> Result<Option<Hash>, Error> {
-        Ok(sqlx::query(
-            r#"
-            SELECT root FROM identities WHERE status = $1 ORDER BY id DESC LIMIT 1
-            "#,
-        )
-        .bind(<&str>::from(status))
-        .fetch_optional(self)
-        .await?
-        .map(|r| r.get::<Hash, _>(0)))
-    }
-
-    async fn get_root_state(self, root: &Hash) -> Result<Option<RootItem>, Error> {
-        // This tries really hard to do everything in one query to prevent race
-        // conditions.
-        Ok(sqlx::query_as::<_, RootItem>(
-            r#"
-            SELECT
-                root,
-                status,
-                pending_as_of as pending_valid_as_of,
-                mined_at as mined_valid_as_of
-            FROM identities
-            WHERE root = $1
-            ORDER BY id
-            LIMIT 1
-            "#,
-        )
-        .bind(root)
-        .fetch_optional(self)
-        .await?)
-    }
-
-    async fn get_latest_insertion_timestamp(self) -> Result<Option<DateTime<Utc>>, Error> {
-        let query = sqlx::query(
-            r#"
-            SELECT insertion_timestamp
-            FROM latest_insertion_timestamp
-            WHERE Lock = 'X';"#,
-        );
-
-        let row = self.fetch_optional(query).await?;
-
-        Ok(row.map(|r| r.get::<DateTime<Utc>, _>(0)))
-    }
-
-    async fn count_unprocessed_identities(self) -> Result<i32, Error> {
-        let query = sqlx::query(
-            r#"
-            SELECT COUNT(*) as unprocessed
-            FROM unprocessed_identities
-            "#,
-        );
-        let result = self.fetch_one(query).await?;
-        Ok(result.get::<i64, _>(0) as i32)
-    }
-
-    async fn count_pending_identities(self) -> Result<i32, Error> {
-        let query = sqlx::query(
-            r#"
-            SELECT COUNT(*) as pending
-            FROM identities
-            WHERE status = $1
-            "#,
-        )
-        .bind(<&str>::from(ProcessedStatus::Pending));
-        let result = self.fetch_one(query).await?;
-        Ok(result.get::<i64, _>(0) as i32)
-    }
-
-    async fn get_provers(self) -> Result<HashSet<ProverConfig>, Error> {
-        Ok(sqlx::query_as(
-            r#"
-            SELECT batch_size, url, timeout_s, prover_type
-            FROM provers
-            "#,
-        )
-        .fetch_all(self)
-        .await?
-        .into_iter()
-        .collect())
-    }
-    async fn insert_prover_configuration(
-        self,
-        batch_size: usize,
-        url: impl ToString,
-        timeout_seconds: u64,
-        prover_type: ProverType,
-    ) -> Result<(), Error> {
-        let url = url.to_string();
-
-        let query = sqlx::query(
-            r#"
-                INSERT INTO provers (batch_size, url, timeout_s, prover_type)
-                VALUES ($1, $2, $3, $4)
-            "#,
-        )
-        .bind(batch_size as i64)
-        .bind(url)
-        .bind(timeout_seconds as i64)
-        .bind(prover_type);
-
-        self.execute(query).await?;
-
-        Ok(())
-    }
-
-    async fn insert_provers(self, provers: HashSet<ProverConfig>) -> Result<(), Error> {
-        if provers.is_empty() {
-            return Ok(());
-        }
-
-        let mut query_builder = sqlx::QueryBuilder::new(
-            r#"
-                  INSERT INTO provers (batch_size, url, timeout_s, prover_type)
-            "#,
-        );
-
-        query_builder.push_values(provers, |mut b, prover| {
-            b.push_bind(prover.batch_size as i64)
-                .push_bind(prover.url)
-                .push_bind(prover.timeout_s as i64)
-                .push_bind(prover.prover_type);
-        });
-
-        let query = query_builder.build();
-
-        self.execute(query).await?;
-        Ok(())
-    }
-
-    async fn remove_prover(self, batch_size: usize, prover_type: ProverType) -> Result<(), Error> {
-        let query = sqlx::query(
-            r#"
-              DELETE FROM provers WHERE batch_size = $1 AND prover_type = $2
-            "#,
-        )
-        .bind(batch_size as i64)
-        .bind(prover_type);
-
-        self.execute(query).await?;
-
-        Ok(())
-    }
-
-    async fn insert_new_identity(
-        self,
-        identity: Hash,
-        eligibility_timestamp: sqlx::types::chrono::DateTime<Utc>,
-    ) -> Result<Hash, Error> {
-        let query = sqlx::query(
-            r#"
-            INSERT INTO unprocessed_identities (commitment, status, created_at, eligibility)
-            VALUES ($1, $2, CURRENT_TIMESTAMP, $3)
-            "#,
-        )
-        .bind(identity)
-        .bind(<&str>::from(UnprocessedStatus::New))
-        .bind(eligibility_timestamp);
-
-        self.execute(query).await?;
-        Ok(identity)
-    }
-
-    async fn insert_new_recovery(
-        self,
-        existing_commitment: &Hash,
-        new_commitment: &Hash,
-    ) -> Result<(), Error> {
-        let query = sqlx::query(
-            r#"
-            INSERT INTO recoveries (existing_commitment, new_commitment)
-            VALUES ($1, $2)
-            "#,
-        )
-        .bind(existing_commitment)
-        .bind(new_commitment);
-        self.execute(query).await?;
-        Ok(())
-    }
-
-    async fn get_latest_deletion(self) -> Result<LatestDeletionEntry, Error> {
-        let query =
-            sqlx::query("SELECT deletion_timestamp FROM latest_deletion_root WHERE Lock = 'X';");
-
-        let row = self.fetch_optional(query).await?;
-
-        if let Some(row) = row {
-            Ok(LatestDeletionEntry {
-                timestamp: row.get(0),
-            })
-        } else {
-            Ok(LatestDeletionEntry {
-                timestamp: Utc::now(),
-            })
-        }
-    }
-
-    async fn update_latest_insertion_timestamp(
-        self,
-        insertion_timestamp: DateTime<Utc>,
-    ) -> Result<(), Error> {
-        let query = sqlx::query(
-            r#"
-            INSERT INTO latest_insertion_timestamp (Lock, insertion_timestamp)
-            VALUES ('X', $1)
-            ON CONFLICT (Lock)
-            DO UPDATE SET insertion_timestamp = EXCLUDED.insertion_timestamp;
-            "#,
-        )
-        .bind(insertion_timestamp);
-
-        self.execute(query).await?;
-        Ok(())
-    }
-
-    async fn update_latest_deletion(self, deletion_timestamp: DateTime<Utc>) -> Result<(), Error> {
-        let query = sqlx::query(
-            r#"
-            INSERT INTO latest_deletion_root (Lock, deletion_timestamp)
-            VALUES ('X', $1)
-            ON CONFLICT (Lock)
-            DO UPDATE SET deletion_timestamp = EXCLUDED.deletion_timestamp;
-            "#,
-        )
-        .bind(deletion_timestamp);
-
-        self.execute(query).await?;
-        Ok(())
-    }
-
-    async fn get_all_recoveries(self) -> Result<Vec<RecoveryEntry>, Error> {
-        Ok(
-            sqlx::query_as::<_, RecoveryEntry>("SELECT * FROM recoveries")
-                .fetch_all(self)
-                .await?,
-        )
-    }
-
-    async fn delete_recoveries<I: IntoIterator<Item = T>, T: Into<U256>>(
-        self,
-        prev_commits: I,
-    ) -> Result<Vec<RecoveryEntry>, Error> {
-        // TODO: upstream PgHasArrayType impl to ruint
-        let prev_commits = prev_commits
-            .into_iter()
-            .map(|c| c.into().to_be_bytes())
-            .collect::<Vec<[u8; 32]>>();
-
-        let res = sqlx::query_as::<_, RecoveryEntry>(
-            r#"
-            DELETE
-            FROM recoveries
-            WHERE existing_commitment = ANY($1)
-            RETURNING *
-            "#,
-        )
-        .bind(&prev_commits)
-        .fetch_all(self)
-        .await?;
-
-        Ok(res)
-    }
-
-    async fn insert_new_deletion(self, leaf_index: usize, identity: &Hash) -> Result<(), Error> {
-        let query = sqlx::query(
-            r#"
-            INSERT INTO deletions (leaf_index, commitment)
-            VALUES ($1, $2)
-            "#,
-        )
-        .bind(leaf_index as i64)
-        .bind(identity);
-
-        self.execute(query).await?;
-        Ok(())
-    }
-
-    // TODO: consider using a larger value than i64 for leaf index, ruint should
-    // have postgres compatibility for u256
-    async fn get_deletions(self) -> Result<Vec<DeletionEntry>, Error> {
-        let query = sqlx::query(
-            r#"
-            SELECT *
-            FROM deletions
-            "#,
-        );
-
-        let result = self.fetch_all(query).await?;
-
-        Ok(result
-            .into_iter()
-            .map(|row| DeletionEntry {
-                leaf_index: row.get::<i64, _>(0) as usize,
-                commitment: row.get::<Hash, _>(1),
-            })
-            .collect::<Vec<DeletionEntry>>())
-    }
-
-    /// Remove a list of entries from the deletions table
-    async fn remove_deletions(self, commitments: &[Hash]) -> Result<(), Error> {
-        let commitments = commitments
-            .iter()
-            .map(|c| c.to_be_bytes())
-            .collect::<Vec<[u8; 32]>>();
-
-        sqlx::query("DELETE FROM deletions WHERE commitment = Any($1)")
-            .bind(commitments)
-            .execute(self)
-            .await?;
-
-        Ok(())
-    }
-
-    async fn get_eligible_unprocessed_commitments(
-        self,
-        status: UnprocessedStatus,
-    ) -> Result<Vec<types::UnprocessedCommitment>, Error> {
-        let query = sqlx::query(
-            r#"
-                SELECT * FROM unprocessed_identities
-                WHERE status = $1 AND CURRENT_TIMESTAMP > eligibility
-                LIMIT $2
-            "#,
-        )
-        .bind(<&str>::from(status))
-        .bind(MAX_UNPROCESSED_FETCH_COUNT);
-
-        let result = self.fetch_all(query).await?;
-
-        Ok(result
-            .into_iter()
-            .map(|row| types::UnprocessedCommitment {
-                commitment: row.get::<Hash, _>(0),
-                status,
-                created_at: row.get::<_, _>(2),
-                processed_at: row.get::<_, _>(3),
-                error_message: row.get::<_, _>(4),
-                eligibility_timestamp: row.get::<_, _>(5),
-            })
-            .collect::<Vec<_>>())
-    }
-
-    async fn get_unprocessed_commit_status(
-        self,
-        commitment: &Hash,
-    ) -> Result<Option<(UnprocessedStatus, String)>, Error> {
-        let query = sqlx::query(
-            r#"
-                SELECT status, error_message FROM unprocessed_identities WHERE commitment = $1
-            "#,
-        )
-        .bind(commitment);
-
-        let result = self.fetch_optional(query).await?;
-
-        if let Some(row) = result {
-            return Ok(Some((
-                row.get::<&str, _>(0).parse().expect("couldn't read status"),
-                row.get::<Option<String>, _>(1).unwrap_or_default(),
-            )));
-        };
-        Ok(None)
-    }
-
-    async fn remove_unprocessed_identity(self, commitment: &Hash) -> Result<(), Error> {
-        let query = sqlx::query(
-            r#"
-                DELETE FROM unprocessed_identities WHERE commitment = $1
-            "#,
-        )
-        .bind(commitment);
-
-        self.execute(query).await?;
-
-        Ok(())
-    }
-
-    async fn identity_exists(self, commitment: Hash) -> Result<bool, Error> {
-        Ok(sqlx::query(
-            r#"
-            select
-            EXISTS (select commitment from unprocessed_identities where commitment = $1) OR
-            EXISTS (select commitment from identities where commitment = $1);
-            "#,
-        )
-        .bind(commitment)
-        .fetch_one(self)
-        .await?
-        .get::<bool, _>(0))
-    }
-
-    // TODO: add docs
-    async fn identity_is_queued_for_deletion(self, commitment: &Hash) -> Result<bool, Error> {
-        let query_queued_deletion =
-            sqlx::query(r#"SELECT exists(SELECT 1 FROM deletions where commitment = $1)"#)
-                .bind(commitment);
-        let row_unprocessed = self.fetch_one(query_queued_deletion).await?;
-        Ok(row_unprocessed.get::<bool, _>(0))
-    }
 }
 
 #[derive(Debug, Error)]
@@ -867,7 +163,7 @@ mod test {
 
     use super::Database;
     use crate::config::DatabaseConfig;
-    use crate::database::DatabaseExt as _;
+    use crate::database::query::DatabaseQuery;
     use crate::identity_tree::{Hash, ProcessedStatus, Status, UnprocessedStatus};
     use crate::prover::{ProverConfig, ProverType};
     use crate::utils::secret::SecretUrl;
@@ -1369,7 +665,7 @@ mod test {
                 .context("Inserting identity")?;
         }
 
-        db.mark_root_as_processed(&roots[2]).await?;
+        db.mark_root_as_processed_tx(&roots[2]).await?;
 
         db.mark_all_as_pending().await?;
 
@@ -1399,7 +695,7 @@ mod test {
                 .context("Inserting identity")?;
         }
 
-        db.mark_root_as_processed(&roots[2]).await?;
+        db.mark_root_as_processed_tx(&roots[2]).await?;
 
         for root in roots.iter().take(3) {
             let root = db
@@ -1442,7 +738,7 @@ mod test {
                 .context("Inserting identity")?;
         }
 
-        db.mark_root_as_mined(&roots[2]).await?;
+        db.mark_root_as_mined_tx(&roots[2]).await?;
 
         for root in roots.iter().take(3) {
             let root = db
@@ -1488,27 +784,27 @@ mod test {
         }
 
         println!("Marking roots up to 2nd as processed");
-        db.mark_root_as_processed(&roots[2]).await?;
+        db.mark_root_as_processed_tx(&roots[2]).await?;
 
         assert_roots_are(&db, &roots[..3], ProcessedStatus::Processed).await?;
         assert_roots_are(&db, &roots[3..], ProcessedStatus::Pending).await?;
 
         println!("Marking roots up to 1st as mined");
-        db.mark_root_as_mined(&roots[1]).await?;
+        db.mark_root_as_mined_tx(&roots[1]).await?;
 
         assert_roots_are(&db, &roots[..2], ProcessedStatus::Mined).await?;
         assert_roots_are(&db, &[roots[2]], ProcessedStatus::Processed).await?;
         assert_roots_are(&db, &roots[3..], ProcessedStatus::Pending).await?;
 
         println!("Marking roots up to 4th as processed");
-        db.mark_root_as_processed(&roots[4]).await?;
+        db.mark_root_as_processed_tx(&roots[4]).await?;
 
         assert_roots_are(&db, &roots[..2], ProcessedStatus::Mined).await?;
         assert_roots_are(&db, &roots[2..5], ProcessedStatus::Processed).await?;
         assert_roots_are(&db, &roots[5..], ProcessedStatus::Pending).await?;
 
         println!("Marking all roots as mined");
-        db.mark_root_as_mined(&roots[num_identities - 1]).await?;
+        db.mark_root_as_mined_tx(&roots[num_identities - 1]).await?;
 
         assert_roots_are(&db, &roots, ProcessedStatus::Mined).await?;
 
@@ -1530,10 +826,10 @@ mod test {
         }
 
         // root[2] is somehow erroneously marked as mined
-        db.mark_root_as_processed(&roots[2]).await?;
+        db.mark_root_as_processed_tx(&roots[2]).await?;
 
         // Later we correctly mark the previous root as mined
-        db.mark_root_as_processed(&roots[1]).await?;
+        db.mark_root_as_processed_tx(&roots[1]).await?;
 
         for root in roots.iter().take(2) {
             let root = db
@@ -1591,7 +887,7 @@ mod test {
             "Root has not yet been mined"
         );
 
-        db.mark_root_as_processed(&roots[0]).await?;
+        db.mark_root_as_processed_tx(&roots[0]).await?;
 
         let root = db
             .get_root_state(&roots[0])
@@ -1632,7 +928,7 @@ mod test {
                 .context("Inserting identity")?;
         }
 
-        db.mark_root_as_processed(&roots[2]).await?;
+        db.mark_root_as_processed_tx(&roots[2]).await?;
 
         let mined_tree_updates = db
             .get_commitments_by_status(ProcessedStatus::Processed)
@@ -1761,7 +1057,7 @@ mod test {
         db.insert_pending_identity(3, &identities[3], &roots[3])
             .await?;
 
-        db.mark_root_as_processed(&roots[0])
+        db.mark_root_as_processed_tx(&roots[0])
             .await
             .context("Marking root as mined")?;
 
@@ -1925,7 +1221,7 @@ mod test {
 
         db.insert_pending_identity(0, &identities[0], &roots[0])
             .await?;
-        db.mark_root_as_mined(&roots[0]).await?;
+        db.mark_root_as_mined_tx(&roots[0]).await?;
 
         db.insert_new_deletion(0, &identities[0]).await?;
 
@@ -1961,7 +1257,7 @@ mod test {
         db.insert_pending_identity(0, &Hash::ZERO, &roots[1])
             .await?;
 
-        db.mark_root_as_mined(&roots[1]).await?;
+        db.mark_root_as_mined_tx(&roots[1]).await?;
 
         let history = db.get_identity_history_entries(&identities[0]).await?;
 
@@ -2002,7 +1298,7 @@ mod test {
         assert_eq!(history[0].leaf_index, Some(0));
         assert!(!history[0].held_back, "Identity should not be held back");
 
-        db.mark_root_as_mined(&roots[0]).await?;
+        db.mark_root_as_mined_tx(&roots[0]).await?;
 
         let history = db.get_identity_history_entries(&identities[0]).await?;
