+++ src/task_monitor.rs
@@ -8,7 +8,8 @@ use tokio::task::JoinHandle;
 use tracing::{info, instrument, warn};
 
 use crate::app::App;
-use crate::database::{Database, DatabaseExt as _};
+use crate::database::query::DatabaseQuery as _;
+use crate::database::Database;
 
 pub mod tasks;
